Remove include/bpf_helpers.h and include/bpf_helper_defs.h

+++ redbpf-probes/src/maps.rs
@@ -345,10 +345,11 @@ impl StackTrace {
 
     pub unsafe fn stack_id(&mut self, ctx: *mut pt_regs, flag: u64) -> Result<c_int, c_int> {
         let ret = bpf_get_stackid(ctx as _, &mut self.def as *mut _ as _, flag);
+        // TODO return i64
         if ret >= 0 {
-            Ok(ret)
+            Ok(ret as i32)
         } else {
-            Err(ret)
+            Err(ret as i32)
         }
     }
 }
@@ -404,7 +405,8 @@ impl ProgramArray {
     pub unsafe fn tail_call<C>(&mut self, ctx: *mut C, index: u32) -> Result<(), i32> {
         let ret = bpf_tail_call(ctx as *mut _, &mut self.def as *mut _ as *mut c_void, index);
         if ret < 0 {
-            return Err(ret);
+            // TODO return i64
+            return Err(ret as i32);
         }
 
         Ok(())
